Moves `SchdulerConfig` to `fthmc/config.py`
